Refactor package folders
